# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bhojnalaya API",
        "description": "REST API documentation for tables, menu items, orders, and order items.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/order-items": {
            "get": {
                "tags": [
                    "Order Items"
                ],
                "summary": "List order items",
                "operationId": "00ef4347ae4117974e08654982418127",
                "responses": {
                    "200": {
                        "description": "Order items fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderItem"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Create an order item",
                "operationId": "ebd626997709bc1943529afd1010f1b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItem"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order-items/{order_item}": {
            "get": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Get a single order item",
                "operationId": "8699abbd250562bdbe21556f605a5692",
                "parameters": [
                    {
                        "name": "order_item",
                        "in": "path",
                        "description": "Order item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order item fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItem"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Update an order item",
                "operationId": "11e15ef68dc964e4a346861bf5b3232d",
                "parameters": [
                    {
                        "name": "order_item",
                        "in": "path",
                        "description": "Order item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItem"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Delete an order item",
                "operationId": "2441e7622348886703617368afc2ea0f",
                "parameters": [
                    {
                        "name": "order_item",
                        "in": "path",
                        "description": "Order item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order item deleted successfully"
                    }
                }
            }
        },
        "/api/tables": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "List dining tables",
                "operationId": "ff9d243a2aac0c783b0b20f5bd9d5913",
                "responses": {
                    "200": {
                        "description": "Dining tables fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Table"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Create a dining table",
                "operationId": "ee3cdb4f76abd42e86f037b17e4d0206",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TableRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dining table created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Table"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tables/{table}": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Get a single dining table",
                "operationId": "53507a0855b5e513cfccea8c5a56c17c",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "description": "Table ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dining table fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Table"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tables"
                ],
                "summary": "Update a dining table",
                "operationId": "d598b5144dac0c77d3ef41d20192669a",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "description": "Table ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TableRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dining table updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Table"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tables"
                ],
                "summary": "Delete a dining table",
                "operationId": "a3c24e6b24285c8d8cbe4bf85db5e816",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "description": "Table ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Dining table deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Table": {
                "required": [
                    "id",
                    "name",
                    "capacity",
                    "status",
                    "is_active"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "T1"
                    },
                    "floor": {
                        "type": "string",
                        "example": "Ground Floor",
                        "nullable": true
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "example": "blank"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "TableRequest": {
                "required": [
                    "name",
                    "capacity",
                    "status"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "T1"
                    },
                    "floor": {
                        "type": "string",
                        "example": "Ground Floor"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "example": "blank"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "OrderItem": {
                "required": [
                    "id",
                    "order_id",
                    "menu_item_id",
                    "quantity",
                    "unit_price",
                    "line_total"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "menu_item_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 220
                    },
                    "line_total": {
                        "type": "number",
                        "format": "float",
                        "example": 440
                    },
                    "notes": {
                        "type": "string",
                        "example": "Less spicy",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:15:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:15:00.000000Z"
                    }
                },
                "type": "object"
            },
            "OrderItemRequest": {
                "required": [
                    "order_id",
                    "menu_item_id",
                    "quantity"
                ],
                "properties": {
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "menu_item_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "notes": {
                        "type": "string",
                        "example": "Less spicy"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The name field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
...
}